feat: new fuzzer for fuzzing multiple types at once * added fuzz target 'crypt2_load_ondisk_fuzz' that tries to load fuzz input as LUKS1, FileVault@, BitLocker, LUKS2 in that order. * added dictionary for this fuzz target * added fuzz target to relevant files

+++ tests/fuzz/oss-fuzz-build.sh
@@ -29,7 +29,7 @@ export CFLAGS="${CFLAGS:-$flags} -I$DEPS_PATH/include"
 export CXXFLAGS="${CXXFLAGS:-$flags} -I$DEPS_PATH/include"
 export LDFLAGS="${LDFLAGS-} -L$DEPS_PATH/lib"
 
-ENABLED_FUZZERS=${ENABLED_FUZZERS:-crypt2_load_fuzz crypt2_load_proto_plain_json_fuzz}
+ENABLED_FUZZERS=${ENABLED_FUZZERS:-crypt2_load_fuzz crypt2_load_ondisk_fuzz crypt2_load_proto_plain_json_fuzz}
 
 mkdir -p $SRC
 mkdir -p $OUT
@@ -49,7 +49,7 @@ in_oss_fuzz && apt-get update && apt-get install -y \
 [ ! -d libprotobuf-mutator ] && git clone --depth 1 https://github.com/google/libprotobuf-mutator.git
 [ ! -d openssl ]    && git clone --depth 1 https://github.com/openssl/openssl
 [ ! -d util-linux ] && git clone --depth 1 https://github.com/util-linux/util-linux
-[ ! -d cryptsetup_fuzzing ] && git clone --depth 1 https://gitlab.com/cryptsetup/cryptsetup_fuzzing.git
+[ ! -d cryptsetup_fuzzing ] && git clone --depth 1 -b multiple_types https://gitlab.com/xflord/cryptsetup_fuzzing.git
 
 cd openssl
 ./Configure --prefix="$DEPS_PATH" --libdir=lib no-shared no-module no-asm
@@ -133,7 +133,9 @@ make -j fuzz-targets
 
 for fuzzer in $ENABLED_FUZZERS; do
     cp tests/fuzz/$fuzzer $OUT
-    cp $SRC/cryptsetup_fuzzing/${fuzzer}_seed_corpus.zip $OUT
+    if [ -e $SRC/cryptsetup_fuzzing/${fuzzer}_seed_corpus.zip ]; then
+      cp $SRC/cryptsetup_fuzzing/${fuzzer}_seed_corpus.zip $OUT
+    fi
 
     # optionally copy the dictionary if it exists
     if [ -e tests/fuzz/${fuzzer}.dict ]; then
